% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{adjMatCGrProb}
\alias{adjMatCGrProb}
\title{Builds the adjacency matrix with probabilities}
\usage{
adjMatCGrProb(data_top, gene_list)
}
\arguments{
\item{data_top}{topology of the full network}

\item{gene_list}{gene clustering output}
}
\value{
adj_cgr: adjacency matrix with probabilities
}
\description{
Builds the adjacency matrix with probabilities
}
