% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{centMedVarCutDistPerc}
\alias{centMedVarCutDistPerc}
\title{Identify the center and radius of each model cluster}
\usage{
centMedVarCutDistPerc(data, clusterRef, percThr = 0.01)
}
\arguments{
\item{data}{gene expression matrix}

\item{clusterRef}{cluster indices of all models}

\item{percThr}{Threshold of permutation test}
}
\value{
: a list of (center, variance, radius(centroid & medoid)) for each cluster
}
\description{
Identify the center and radius of each model cluster
}
