% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{convAdjTop}
\alias{convAdjTop}
\title{Converting the adjacency matrix of a gene circuit to the topology file (Source/Target/Interaction Type)}
\usage{
convAdjTop(adjMatL, numbG, colN)
}
\arguments{
\item{adjMatL:}{adjacency matrix}

\item{numbG:}{the number of genes}

\item{colN:}{a vector of gene names}
}
\value{
data_cgr_top:  a data frame of the circuit topology
}
\description{
Converting the adjacency matrix of a gene circuit to the topology file (Source/Target/Interaction Type)
}
