% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{convTopAdj}
\alias{convTopAdj}
\title{Converting the topology file (Source/Target/Interaction Type) of a gene circuit to the adjacency matrix d}
\usage{
convTopAdj(data_t, numbG, colN)
}
\arguments{
\item{data_t:}{a data frame of the circuit topology}

\item{numbG:}{the number of genes}

\item{colN:}{a vector of gene names}
}
\value{
adjMatL: adjacency matrix
}
\description{
Converting the topology file (Source/Target/Interaction Type) of a gene circuit to the adjacency matrix d
}
