% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{difLen}
\alias{difLen}
\title{Mismatches between the sampled and reference circuit topologies}
\usage{
difLen(samp_top, ref_top)
}
\arguments{
\item{sam_top:}{Adjacency matrix of the sampled circuit topology}

\item{ref_top:}{Adjacency matrix of the reference circuit topology}
}
\value{
: number of mismatching edges
}
\description{
Mismatches between the sampled and reference circuit topologies
}
