% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{find_in_list}
\alias{find_in_list}
\title{Find an element in a list}
\usage{
find_in_list(xli, n, elem)
}
\arguments{
\item{xli}{the list}

\item{n}{the length of the list}

\item{elem}{the element to search}
}
\value{
indli: the index of the found element in the list, or 0 if not found
}
\description{
Find an element in a list
}
