% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{find_inout}
\alias{find_inout}
\title{Identify input and output nodes from a circuit topology}
\usage{
find_inout(adjMatrPr, numb_cgnodes)
}
\arguments{
\item{adjMatrPr:}{Adjacency matrix with probabilities of CG-circuits derived from the full network}

\item{num_cgnodes:}{number of CG nodes (an output of gene clustering)}
}
\value{
resNod: a list of input nodes (1) and output nodes (2)
}
\description{
Input nodes: nodes with only outward edges; Output nodes: nodes with only inward edges
}
