% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{find_topology}
\alias{find_topology}
\title{Check whether the circuit topology has been sampled or not}
\usage{
find_topology(top_list, samp_top, minN, maxN)
}
\arguments{
\item{top_list}{a list of sampled topologies}

\item{samp_top}{the current topology}

\item{minN}{the range of circuit topologies in the list to be searched (minimum index value)}

\item{maxN}{the range of circuit topologies in the list to be searched (maximum index value)}
}
\value{
i: the index of the found topology in the list, or 0 if not found
}
\description{
Check whether the circuit topology has been sampled or not
}
