% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{gaInitial_gen}
\alias{gaInitial_gen}
\title{Generating a series of randomly generated initial CG circuit topologies}
\usage{
gaInitial_gen(circuit_top, gene_list, numbNewTop)
}
\arguments{
\item{circuit_top:}{the original circuit topology}

\item{gene_list:}{gene clustering output}

\item{numbNewTop:}{number of new circuit topologies to be generated}
}
\value{
resM: a list of randomly generated initial CG circuit topologies
}
\description{
Generating a series of randomly generated initial CG circuit topologies
}
