% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gen_heatmap_hca}
\alias{gen_heatmap_hca}
\title{Generating HCA Heatmap}
\usage{
gen_heatmap_hca(logscData)
}
\arguments{
\item{logscData:}{simulated data after log scaling and standardization}
}
\description{
Generating HCA Heatmap
}
