% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gen_pca_plot}
\alias{gen_pca_plot}
\title{Generating PCA Scatterplot}
\usage{
gen_pca_plot(logscData)
}
\arguments{
\item{logscData:}{simulated data after log scaling and standardization}
}
\description{
Generating PCA Scatterplot
}
