% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gen_RACIPE}
\alias{gen_RACIPE}
\title{RACIPE simulations}
\usage{
gen_RACIPE(dftop, nModels, integrateStepSize = 0.02, simulationTime = 200)
}
\arguments{
\item{dftop:}{circuit topology}

\item{nModels:}{number of RACIPE models generated}

\item{integrateStepSize:}{step size for the ODE integration}

\item{simulationTime:}{simulation time of ODE for each RACIPE model}
}
\value{
logscData: simulated data after log scaling and standardization
}
\description{
RACIPE simulations
}
