% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{geneClustInd}
\alias{geneClustInd}
\title{GENE CLUSTERING BY INDIVIDUAL MODELS (HCA)}
\usage{
geneClustInd(logscData, numbGeneClust)
}
\arguments{
\item{logscData:}{simulated data after log scaling and standardization}

\item{numbGeneClust:}{number of gene clusters}
}
\value{
gene_list: gene clustering output
}
\description{
GENE CLUSTERING BY INDIVIDUAL MODELS (HCA)
}
