% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{geneClustMedian}
\alias{geneClustMedian}
\title{GENE CLUSTERING BY MEDIAN VALUES}
\usage{
geneClustMedian(clustData, clSize, numbGeneClust)
}
\arguments{
\item{clustData:}{model clustering data}

\item{clSize:}{size of each model clusters (an output of the model clustering results)}

\item{numbGeneClust:}{number of gene clusters}
}
\value{
gene_list: gene clustering output
}
\description{
GENE CLUSTERING BY MEDIAN VALUES
}
