% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{initialize_topology}
\alias{initialize_topology}
\title{Build the most dense circuit topology (adj. matrix in a vector``)}
\usage{
initialize_topology(adj_matrProb, numb_cgnodes)
}
\arguments{
\item{adj_matrProb}{adjacency matrix with probabilities}

\item{numb_cgnodes}{the number of CG nodes}
}
\value{
adj_Matr:  the adj. matrix of the most dense topology as a vector
}
\description{
Build the most dense circuit topology (adj. matrix in a vector``)
}
