% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{inNWsimilarityRefCGr}
\alias{inNWsimilarityRefCGr}
\title{Circuit scoring}
\usage{
inNWsimilarityRefCGr(
  dataRow,
  clusterRef,
  cenMedRef,
  cutOffM,
  gene_list,
  inNodes,
  topol_cgr,
  modelsCGr = 10000
)
}
\arguments{
\item{dataRow}{gene expression matrix with gene in rows}

\item{clusterRef}{the cluster indices of all models}

\item{cenMedRef}{cluster centers}

\item{cutOffM}{cluster radii}

\item{gene_list}{gene clustering output}

\item{inNodes}{a list of input nodes}

\item{topol_cgr}{the current CG circuit topology}

\item{modelsCGr}{the number of RACIPE models to be simulated (10000)}
}
\value{
scoresOuts: 1: the total score, 2: the number of noisy models
}
\description{
Circuit scoring
}
