% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{modClustHCA}
\alias{modClustHCA}
\title{Model Clustering By hierarchical clustering analysis (HCA)}
\usage{
modClustHCA(logscData, numbClust)
}
\arguments{
\item{logscData}{simulated data after log scaling and standardization}

\item{numbClust}{number of model clusters}
}
\value{
res: model clustering output: (1: cluster sizes; 2: clustering rearranged data; 3: cluster indices)
}
\description{
Model Clustering By hierarchical clustering analysis (HCA)
}
