% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{modClustKmeans}
\alias{modClustKmeans}
\title{MODEL CLUSTERING BY K-MEANS}
\usage{
modClustKmeans(data, numbClust, clustCenters)
}
\arguments{
\item{data}{data matrix for k-means clustering (either logscData or projected data)}

\item{numbClust}{number of model clusters}

\item{clustCenters}{coordinates of the cluster center}
}
\value{
res: model clustering output
}
\description{
MODEL CLUSTERING BY K-MEANS
}
