% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_combined.R
\name{opt_combined}
\alias{opt_combined}
\title{Circuit optimization (combined MH, MA & TE)}
\usage{
opt_combined(
  network_top,
  data,
  clusterRef,
  cenMedRef,
  cutOffM,
  gene_list,
  inTopsM,
  output = "Results",
  nRepeat = 5,
  nIter = 1400,
  modelsCGr = 10000,
  numbThr = 40,
  nSim = 20
)
}
\arguments{
\item{network_top:}{topology of the full network}

\item{data:}{processed gene expression matrix}

\item{clusterRef:}{cluster indices of all models}

\item{cenMedRef:}{cluster centers}

\item{cutOffM:}{cluster radii}

\item{gene_list:}{gene clustering output}

\item{inTopsM:}{a list of all initial circuit topologies}

\item{output:}{a string of file prefix for saving results ("Results")}

\item{nRepeat:}{number of repeats of RACIPE simulations for each new circuit topology (5)
A new circuit is simulated by RACIPE nRepeat times for robust score evaluations; 
The scores will then be saved and used in future iterations, when the circuits are sampled again.}

\item{nIter:}{number of iterations for each simulation}

\item{modelsCGr:}{number of RACIPE models to be simulated (10000)}

\item{numbThr:}{number of HPC threads}

\item{nSim:}{number of simultaneous simulations}
}
\value{
(in results files)
}
\description{
Circuit optimization (combined MH, MA & TE)
}
