% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_MH_multi.R
\name{opt_MH_multi}
\alias{opt_MH_multi}
\title{Circuit optimization with Metropolis-Hastings (MH) algorithm (multiple threads)}
\usage{
opt_MH_multi(
  network_top,
  data,
  clusterRef,
  cenMedRef,
  cutOffM,
  gene_list,
  inTopsM,
  output = "Results",
  nRepeat = 5,
  nIter = 1400,
  modelsCGr = 10000,
  tempM = 60,
  numbThr = 40,
  nSim = 20
)
}
\arguments{
\item{network_top:}{topology of the full network}

\item{data:}{processed gene expression matrix}

\item{clusterRef:}{cluster indices of all models}

\item{cenMedRef:}{cluster centers}

\item{cutOffM:}{cluster radii}

\item{gene_list:}{gene clustering output}

\item{inTopsM:}{a list of all initial circuit topologies}

\item{output:}{a string of file prefix for saving results ("Results")}

\item{nRepeat:}{number of repeats of RACIPE simulations for each new circuit topology (5)
A new circuit is simulated by RACIPE nRepeat times for robust score evaluation; 
The scores will then be saved and used in future iterations, when the circuits are sampled again.}

\item{nIter:}{number of iterations for each simulation (1400)}

\item{modelsCGr:}{number of RACIPE models to be simulated (10000)}

\item{tempM:}{temperature for MH (60)}

\item{numbThr:}{number of requested threads for HPC (40)}

\item{nSim:}{number of parallel simulations (20)}
}
\value{
df: topology of the optimized CG circuit
}
\description{
Circuit optimization with Metropolis-Hastings (MH) algorithm (multiple threads)
}
