% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_SA.R
\name{opt_SA}
\alias{opt_SA}
\title{Circuit optimization with Simulated Annealing (SA) algorithm}
\usage{
opt_SA(
  network_top,
  data,
  clusterRef,
  cenMedRef,
  cutOffM,
  gene_list,
  init_top,
  output = "Results",
  nRepeat = 5,
  modelsCGr = 10000,
  maxT = 150,
  decayRate1 = 0.8,
  decayRate2 = 0.6,
  threshT = 40,
  iter_per_temp = 100
)
}
\arguments{
\item{network_top:}{topology of the full network}

\item{data:}{processed gene expression matrix}

\item{clusterRef:}{cluster indices of all models}

\item{cenMedRef:}{cluster centers}

\item{cutOffM:}{cluster radii}

\item{gene_list:}{gene clustering output}

\item{init_top:}{initial circuit topology}

\item{output:}{a string of file prefix for saving results ("Results")}

\item{nRepeat:}{number of repeats of RACIPE simulations for each new circuit topology (5)
A new circuit is simulated by RACIPE nRepeat times for robust score evaluation; 
The scores will then be saved and used in future iterations, when the circuits are sampled again.}

\item{modelsCGr:}{number of RACIPE models to be simulated (10000)}

\item{maxT:}{maximum/initial temperature in SA (150)}

\item{threshT:}{a second temperature in SA, below which SA has a slower temperature decaying rate (40)}

\item{decayRate1:}{1st temperature decaying rate (geometrically decaying) (0.8)}

\item{decayRate2:}{2nd temperature decaying rate (0.6), until temperature = 1 (current implementation)}

\item{iter_per_temp:}{number of iterations for each fixed temperature (100)}
}
\value{
df: topology of the optimized CG circuit
}
\description{
Circuit optimization with Simulated Annealing (SA) algorithm
}
