% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{plot_network}
\alias{plot_network}
\title{Visualize network topology}
\usage{
plot_network(tf_links, height = "300px")
}
\arguments{
\item{tf_links}{circuit topology}

\item{height}{plot height ("300px")}
}
\value{
network plot
}
\description{
Visualize network topology
}
