% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reordering}
\alias{reordering}
\title{Reordering the gene clusters in the gene expression matrix}
\usage{
reordering(logscData, gene_list, geneGroupOrder = NULL)
}
\arguments{
\item{logscData}{current gene expression data matrix}

\item{gene_list}{gene clustering output}

\item{geneGroupOrder}{desired order of gene groups}
}
\value{
a list of reordered data (logscData) and an updated gene list (gene_list)
}
\description{
Reordering the gene clusters in the gene expression matrix
}
