% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit.R
\name{signChanges_top}
\alias{signChanges_top}
\title{Making sign changes to some edges of a circuit topology}
\usage{
signChanges_top(
  samp_top,
  numb_genes,
  numb_changes,
  pos_changes,
  inNodes,
  list_top
)
}
\arguments{
\item{samp_top}{the current circuit topology}

\item{numb_genes}{number of genes of the circuit}

\item{numb_changes}{number of edges with sign changes (>0)}

\item{pos_changes}{indices of edges that can have sign changes}

\item{inNodes}{a list of input nodes, output from the function find_inout}

\item{list_top}{list of sampled circuit topologies, we make sure they are not sampled again}
}
\value{
cTop: the modified circuit topology
}
\description{
this process will be tried 3 times if no new circuit topology is sampled
}
